pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "python-yarbo"
version = "0.1.0"
description = "Python library for local and cloud control of Yarbo robot mowers via MQTT"
readme = "README.md"
license = { text = "MIT" }
authors = [{ name = "Markus Lassfolk" }]
keywords = ["yarbo", "mqtt", "robot", "mower", "snow-blower", "iot", "smart-home"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Home Automation",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
requires-python = ">=3.11"
dependencies = [
    "paho-mqtt>=2.0",
    "aiohttp>=3.9",
]

[project.optional-dependencies]
cloud = [
    "cryptography>=42.0",
]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=5.0",
    "ruff>=0.4",
    "mypy>=1.10",
    "bandit[toml]>=1.7",
    "pip-audit>=2.7",
    "pre-commit>=3.7",
    "cryptography>=42.0",
    "types-aiofiles",
]

[project.urls]
Homepage = "https://github.com/markus-lassfolk/python-yarbo"
Repository = "https://github.com/markus-lassfolk/python-yarbo"
"Bug Tracker" = "https://github.com/markus-lassfolk/python-yarbo/issues"
Documentation = "https://github.com/markus-lassfolk/python-yarbo#readme"
Changelog = "https://github.com/markus-lassfolk/python-yarbo/blob/main/CHANGELOG.md"
"yarbo-reversing" = "https://github.com/markus-lassfolk/yarbo-reversing"

[tool.hatch.build.targets.wheel]
packages = ["src/yarbo"]

# ---------------------------------------------------------------------------
# pytest
# ---------------------------------------------------------------------------

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = ["--tb=short", "-q"]

# ---------------------------------------------------------------------------
# coverage
# ---------------------------------------------------------------------------

[tool.coverage.run]
source = ["src/yarbo"]
omit = ["tests/*", "examples/*"]

[tool.coverage.report]
show_missing = true
skip_covered = false
fail_under = 70

# ---------------------------------------------------------------------------
# mypy
# ---------------------------------------------------------------------------

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
exclude = ["tests/", "examples/", "docs/"]

# ---------------------------------------------------------------------------
# bandit
# ---------------------------------------------------------------------------

[tool.bandit]
exclude_dirs = ["tests", "examples", "docs"]
skips = ["B101"]  # assert_used — allowed in tests
